+++ src/utils/index.ts
@@ -1,2 +1,3 @@
+export * from './DeploymentBuilder'
 export * from './Environment'
 export { addModelToFormData } from './Helper'
